refactor(core): move import/export tracking into module_tracking folder
